FTH-129: refactor FaithImageStorageService to use menaS3Client and improve image upload handling
